Large refactoring commit
