+++ pirouette.cabal
@@ -25,19 +25,18 @@ library
       Language.Pirouette.QuasiQuoter
       Language.Pirouette.QuasiQuoter.Syntax
       Language.Pirouette.QuasiQuoter.ToTerm
-      ListT.Extra
       ListT.Weighted
       ListT.Weighted.Examples
       Pirouette
       Pirouette.Monad
       Pirouette.Monad.Maybe
-      Pirouette.Pipeline
+      Pirouette.SMT
       Pirouette.SMT.Base
       Pirouette.SMT.Constraints
       Pirouette.SMT.FromTerm
       Pirouette.SMT.Monadic
       Pirouette.Symbolic.Eval
-      Pirouette.Symbolic.Eval.Helpers
+      Pirouette.Symbolic.Eval.BranchingHelpers
       Pirouette.Symbolic.Eval.SMT
       Pirouette.Symbolic.Eval.Types
       Pirouette.Symbolic.Prover
@@ -48,15 +47,17 @@ library
       Pirouette.Term.Syntax.Pretty.Class
       Pirouette.Term.Syntax.Subst
       Pirouette.Term.Syntax.SystemF
-      Pirouette.Term.Transformations
       Pirouette.Term.TransitiveDeps
       Pirouette.Term.TypeChecker
       Pirouette.Transformations
       Pirouette.Transformations.Contextualize
       Pirouette.Transformations.Defunctionalization
+      Pirouette.Transformations.ElimEvenOddMutRec
       Pirouette.Transformations.EtaExpand
+      Pirouette.Transformations.Inline
       Pirouette.Transformations.Monomorphization
       Pirouette.Transformations.Prenex
+      Pirouette.Transformations.Term
       Pirouette.Transformations.Utils
       PureSMT
       PureSMT.Process
@@ -121,11 +122,11 @@ test-suite spec
       Pirouette.Symbolic.ProveSpec
       Pirouette.Term.Syntax.BaseSpec
       Pirouette.Term.Syntax.SystemFSpec
-      Pirouette.Term.TransformationsSpec
       Pirouette.Transformations.DefunctionalizationSpec
       Pirouette.Transformations.EtaExpandSpec
       Pirouette.Transformations.MonomorphizationSpec
       Pirouette.Transformations.PrenexSpec
+      Pirouette.Transformations.TermSpec
       Paths_pirouette
   autogen-modules:
       Paths_pirouette
