+++ src/Pirouette/SMT.hs
@@ -1,0 +1,7 @@
+-- |Re-exports of Pirouette's wrapper on top of "PureSMT"
+module Pirouette.SMT (module X) where
+
+import Pirouette.SMT.Base as X
+import Pirouette.SMT.Constraints as X
+import Pirouette.SMT.FromTerm as X
+import Pirouette.SMT.Monadic as X
