+++ src/Pirouette/Symbolic/Eval/SMT.hs
@@ -25,9 +25,7 @@ import qualified Data.List as List
 import qualified Data.Map.Strict as M
 import qualified Data.Set as S
 import Pirouette.Monad
-import qualified Pirouette.SMT.Constraints as C
-import Pirouette.SMT.FromTerm
-import qualified Pirouette.SMT.Monadic as SMT
+import Pirouette.SMT hiding (Constraint)
 import Pirouette.Symbolic.Eval.Types
 import Pirouette.Term.Syntax
 import Prettyprinter (align, enclose, vsep, (<+>))
@@ -102,34 +100,34 @@ instance Pretty Model where
       f other = other
 
 -- HACKS DUE TO #106: https://github.com/tweag/pirouette/issues/106
-type HackSolver lang a = SMT.SolverT (ReaderT (PrtOrderedDefs lang) IO) a
+type HackSolver lang a = SolverT (ReaderT (PrtOrderedDefs lang) IO) a
 
-hackSolver :: (MonadIO m) => PureSMT.Solver -> SMT.SolverT m a -> m a
-hackSolver s = flip runReaderT s . SMT.unSolverT
+hackSolver :: (MonadIO m) => PureSMT.Solver -> SolverT m a -> m a
+hackSolver s = flip runReaderT s . unSolverT
 
 hackSolverPrt :: PureSMT.Solver -> PrtOrderedDefs lang -> HackSolver lang a -> IO a
-hackSolverPrt s defs = flip runReaderT defs . flip runReaderT s . SMT.unSolverT
+hackSolverPrt s defs = flip runReaderT defs . flip runReaderT s . unSolverT
 
 -- | Instance necessary to call the 'PureSMT.solve' function.
-instance (SMT.LanguageSMT lang) => PureSMT.Solve lang where
+instance (LanguageSMT lang) => PureSMT.Solve lang where
   type Ctx lang = SolverSharedCtx lang
   type Problem lang = SolverProblem lang
   initSolver SolverSharedCtx {..} s = hackSolver s $ do
-    _ <- runExceptT $ SMT.declareDatatypes solverCtxDatatypes
-    mapM_ (uncurry SMT.declareRawFun) solverCtxUninterpretedFuncs
+    _ <- runExceptT $ declareDatatypes solverCtxDatatypes
+    mapM_ (uncurry declareRawFun) solverCtxUninterpretedFuncs
 
   solveProblem (CheckPath CheckPathProblem {..}) s =
     hackSolverPrt s cpathDefs $ pathIsPlausible cpathState
     where
       pathIsPlausible ::
-        (SMT.LanguageSMT lang) =>
+        (LanguageSMT lang) =>
         SymEvalSt lang ->
         HackSolver lang Bool
       pathIsPlausible env
         | sestValidated env = return True -- We already validated this branch before; nothing new was learnt.
         | otherwise = do
-          SMT.solverPush
-          decl <- runExceptT (SMT.declareVariables (sestGamma env))
+          solverPush
+          decl <- runExceptT (declareVariables (sestGamma env))
           case decl of
             Right _ -> return ()
             Left err -> error err
@@ -138,10 +136,10 @@ instance (SMT.LanguageSMT lang) => PureSMT.Solve lang where
           -- And if a partial translation is already unsat,
           -- so is the translation of the whole set of constraints.
           void $ assertConstraint (sestKnownNames env) (sestConstraint env)
-          res <- SMT.checkSat
-          SMT.solverPop
+          res <- checkSat
+          solverPop
           return $ case res of
-            SMT.Unsat -> False
+            Unsat -> False
             _ -> True
   solveProblem (CheckProperty CheckPropertyProblem {..}) s =
     hackSolverPrt s cpropDefs $ checkProperty cpropOut cpropIn cpropAxioms cpropState
@@ -151,82 +149,82 @@ instance (SMT.LanguageSMT lang) => PureSMT.Solve lang where
       -- This is equivalent to the unsatisfiability of
       -- pathConstraints /\ cOut /\ (not cIn).
       checkProperty ::
-        (SMT.LanguageSMT lang, LanguagePretty lang) =>
+        (LanguageSMT lang, LanguagePretty lang) =>
         Constraint lang ->
         Maybe (Constraint lang) ->
         [UniversalAxiom lang] ->
         SymEvalSt lang ->
         HackSolver lang PruneResult
       checkProperty cOut cIn axioms env = do
-        SMT.solverPush
+        solverPush
         let vars = sestGamma env
-        decl <- runExceptT (SMT.declareVariables vars)
+        decl <- runExceptT (declareVariables vars)
         instantiateAxiomWithVars axioms env
         case decl of
           Right _ -> return ()
           Left err -> error err
         (cstrTotal, _cstrUsedAnyUFs) <- assertConstraint (sestKnownNames env) (sestConstraint env)
         (outTotal, _outUsedAnyUFs) <- assertConstraint (sestKnownNames env) cOut
-        inconsistent <- SMT.checkSat
+        inconsistent <- checkSat
         case (inconsistent, cIn) of
-          (SMT.Unsat, _) -> do
-            SMT.solverPop
+          (Unsat, _) -> do
+            solverPop
             return PruneInconsistentAssumptions
           (_, Nothing) -> do
-            SMT.solverPop
+            solverPop
             return PruneUnknown
           (_, Just cIn') -> do
             (inTotal, _inUsedAnyUFs) <- assertNotConstraint (sestKnownNames env) cIn'
             let everythingWasTranslated = cstrTotal && outTotal && inTotal
             -- Any usedAnyUFs = cstrUsedAnyUFs <> outUsedAnyUFs <> inUsedAnyUFs
-            result <- SMT.checkSat
+            result <- checkSat
             -- liftIO $ print result
             -- liftIO $ print $ pretty (sestConstraint env)
             -- liftIO $ print (cstrTotal, outTotal, inTotal)
             finalResult <- case result of
-              SMT.Unsat -> return PruneImplicationHolds
+              Unsat -> return PruneImplicationHolds
               -- If a partial translation of the constraints is SAT,
               -- it does not guarantee us that the full set of constraints is satisfiable.
               -- Only in the case in which we could translate the entire thing to prove.
-              SMT.Sat | everythingWasTranslated -> do
-                m <- if null vars then pure [] else SMT.getModel (M.keys vars)
+              Sat | everythingWasTranslated -> do
+                m <- if null vars then pure [] else getModel (M.keys vars)
                 pure $ PruneCounterFound (Model m)
               _ -> return PruneUnknown
-            SMT.solverPop
+            solverPop
             return finalResult
 
 assertConstraint,
   assertNotConstraint ::
-    (PirouetteReadDefs lang m, SMT.LanguageSMT lang, SMT.ToSMT meta, MonadIO m) =>
+    (PirouetteReadDefs lang m, LanguageSMT lang, MonadIO m) =>
     S.Set Name ->
-    C.Constraint lang meta ->
-    SMT.SolverT m (Bool, UsedAnyUFs)
-assertConstraint knownNames c@C.Bot = do
-  (done, usedAnyUFs, expr) <- C.constraintToSExpr knownNames c
-  SMT.assert expr
+    Constraint lang ->
+    SolverT m (Bool, UsedAnyUFs)
+assertConstraint knownNames c@Bot = do
+  (done, usedAnyUFs, expr) <- constraintToSExpr knownNames c
+  assert expr
   pure (done, usedAnyUFs)
-assertConstraint knownNames (C.And atomics) = do
+assertConstraint knownNames (And atomics) = do
   (dones, usedAnyUFs) <-
     unzip
       <$> forM
         atomics
         ( \atomic -> do
             -- do it one by one
-            (done, usedAnyUFs, expr) <- C.constraintToSExpr knownNames (C.And [atomic])
-            SMT.assert expr
+            (done, usedAnyUFs, expr) <- constraintToSExpr knownNames (And [atomic])
+            assert expr
             pure (done, usedAnyUFs)
         )
   pure (and dones, mconcat usedAnyUFs)
 assertNotConstraint knownNames c = do
-  (done, usedAnyUFs, expr) <- C.constraintToSExpr knownNames c
-  SMT.assertNot expr
+  (done, usedAnyUFs, expr) <- constraintToSExpr knownNames c
+  assertNot expr
   pure (done, usedAnyUFs)
 
 -- TODO: why is this needed, what needs to be done on the TODO below?
 instantiateAxiomWithVars ::
-  (SMT.LanguageSMT lang, MonadIO m) => [UniversalAxiom lang] -> SymEvalSt lang -> SMT.SolverT m ()
+  (LanguageSMT lang, MonadIO m) => [UniversalAxiom lang] -> SymEvalSt lang -> SolverT m ()
 instantiateAxiomWithVars axioms env =
-  SMT.SolverT $ do
+  SolverT $ do
     solver <- ask
     liftIO $
       mapM_
@@ -235,7 +233,7 @@ instantiateAxiomWithVars axioms env =
               ( \UniversalAxiom {..} ->
                   when (List.any (\(_, tyV) -> tyV == ty) boundVars) $
                     if length boundVars == 1
-                      then void $ PureSMT.assert solver (axiomBody [PureSMT.symbol (SMT.toSmtName name)])
+                      then void $ PureSMT.assert solver (axiomBody [PureSMT.symbol (toSmtName name)])
                       else error "Several universally bound variables not handled yet" -- TODO
               )
               axioms
