+++ tests/unit/Language/Pirouette/PlutusIR/ToTermSpec.hs
@@ -10,7 +10,7 @@ import Debug.Trace
 import Language.Pirouette.PlutusIR.ToTerm
 import Pirouette.Monad
 import Pirouette.Term.Syntax
-import Pirouette.Term.Transformations
+import Pirouette.Transformations.Term
 import Test.Tasty
 import Test.Tasty.HUnit
